save 0222

+++ src/Controller/AuthorController.php
@@ -15,15 +15,26 @@ use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
 use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Symfony\Contracts\Cache\ItemInterface;
+use Symfony\Contracts\Cache\TagAwareCacheInterface;
 
 class AuthorController extends AbstractController
 {
     #[Route('api/authors', name: 'app_author', methods: ['GET'])]
 
-    public function getAllAuthors(AuthorRepository $authorRepository, SerializerInterface $serializer): JsonResponse
+    public function getAllAuthors(AuthorRepository $authorRepository, SerializerInterface $serializer, Request $request, TagAwareCacheInterface $cache): JsonResponse
     {
-        $authorList = $authorRepository->findAll();
-        $jsonAuthorList = $serializer->serialize($authorList, 'json', ['groups' => 'getAuthors']);
+        $page = $request->get('page', 1);
+        $limit = $request->get('limit', 5);
+        $idCache = 'getAllAuthors' . $page . '-' . $limit;
+        // $authorList = $authorRepository->findAll();
+        // note: Désactiver lazy loading est moins efficace que le gérer. Voir BookController et BookRepository pour la meilleure manière de faire
+        $jsonAuthorList = $cache->get($idCache, function (ItemInterface $item) use ($authorRepository, $page, $limit, $serializer) {
+            $item->tag('authorsCache');
+            $item->expiresAfter(30*60);
+            $authorList = $authorRepository->findAllWithPagination($page, $limit);
+            return $serializer->serialize($authorList, 'json', ['groups' => 'getAuthors']);
+        });
         return new JsonResponse($jsonAuthorList, Response::HTTP_OK, [], true);
     }
 
@@ -31,6 +42,7 @@ class AuthorController extends AbstractController
 
     public function getDetailAuthor(int $id, SerializerInterface $serializer, AuthorRepository $authorRepository): JsonResponse
     {
+
         $author = $authorRepository->find($id);
         if ($author) {
             $jsonAuthor = $serializer->serialize($author, 'json', ['groups' => 'getAuthors']);
@@ -41,8 +53,10 @@ class AuthorController extends AbstractController
 
     #[Route('api/authors/{id}', name: 'deleteAuthor', methods: ['DELETE'])]
 
-    public function deleteAuthor(int $id, AuthorRepository $authorRepository, EntityManagerInterface $em): JsonResponse
+    public function deleteAuthor(int $id, AuthorRepository $authorRepository, EntityManagerInterface $em, TagAwareCacheInterface $cache): JsonResponse
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, "Vous n'avez pas les autorisations nécessaires pour supprimer un Auteur");
+        $cache->invalidateTags(['authorsCache']);
         $author = $authorRepository->find($id);
         $em->remove($author);
         $em->flush();
@@ -51,9 +65,10 @@ class AuthorController extends AbstractController
 
     #[route('api/authors', name: 'createAuthor', methods: ['POST'])]
 
-    public function createAuthor(Request $request, SerializerInterface $serializer, EntityManagerInterface $em, UrlGeneratorInterface $urlGenerator, BookRepository $bookRepository,
-    ValidatorInterface $validator): JsonResponse
+    public function createAuthor(Request $request, SerializerInterface $serializer, EntityManagerInterface $em, UrlGeneratorInterface $urlGenerator, BookRepository $bookRepository, ValidatorInterface $validator, TagAwareCacheInterface $cache): JsonResponse
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, "Vous n'avez pas les autorisations nécessaires pour ajouter un Auteur");
+        $cache->invalidateTags(['authorsCache']);
         $author = $serializer->deserialize($request->getContent(), Author::class, 'json');
         $content = $request->toArray();
         $idBooks = $content['idBook'] ?? -1;
@@ -80,8 +95,10 @@ class AuthorController extends AbstractController
 
     #[Route('api/authors/{id}', name: 'updateAuthor', methods: ['PUT'])]
 
-    public function updateAuthor(int $id, Request $request, SerializerInterface $serializer, AuthorRepository $authorRepository, EntityManagerInterface $em, BookRepository $bookRepository): JsonResponse
+    public function updateAuthor(int $id, Request $request, SerializerInterface $serializer, AuthorRepository $authorRepository, EntityManagerInterface $em, BookRepository $bookRepository, TagAwareCacheInterface $cache): JsonResponse
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, "Vous n'avez pas les autorisations nécessaires pour modifier un Auteur");
+        $cache->invalidateTags(['authorsCache']);
         $currentAuthor = $authorRepository->find($id);
         $updateAuthor = $serializer->deserialize(
             $request->getContent(),
